+++ paws.common/tests/testthat/test_custom_s3.R
@@ -312,7 +312,7 @@ test_that("ignore redirect if already redirected", {
   expect_equal(actual, req)
 })
 
-test_that("ignore redirect if unable to find S3 region", {
+test_that("default to head_bucket for final region check", {
   raw_error <- charToRaw(paste0(
     "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<Error><Code>PermanentRedirect</Code>",
     "<Message>Dummy Error</Message><Endpoint>foo.s3.us-east-2.amazonaws.com</Endpoint>",
@@ -323,11 +323,18 @@ test_that("ignore redirect if unable to find S3 region", {
     status_code = 301,
     body = raw_error
   )
-  actual <- s3_redirect_from_error(req)
-  expect_equal(actual, req)
-  expect_equal(actual$http_response$body, raw_error)
-})
+  mock_head_bucket <- mock2(list(BucketRegion = "bar"))
+  mock_s3 <- mock2(list(head_bucket = mock_head_bucket))
+  mockery::stub(s3_get_bucket_region, "s3", mock_s3)
 
+  error <- decode_xml(raw_error)$Error
+
+  actual <- s3_get_bucket_region(req, error, "foo")
+
+  head_bucket_args <- mockery::mock_args(mock_head_bucket)[[1]]
+  expect_equal(head_bucket_args, list(Bucket = "foo"))
+  expect_equal(actual, "bar")
+})
 
 test_that("redirect request from http response error", {
   req <- build_request(bucket = "foo", operation = "ListObjects")